egglog_bridge

Module proof_format

A proof format for egglog programs, based on the Rocq format and checker from Tia Vu, Ryan Doegens, and Oliver Flatt.

Structs

CongProof

EqProofId

an id identifying proofs of equality between two terms within a ProofStore

PrettyPrintConfig

ProofStore

A hash-cons store for proofs and terms related to an egglog program.

TermDag

TermId

an id identifying terms within a TermDag

TermProofId

an id identifying proofs of terms within a ProofStore

Enums

EqProof

Premise

Term

TermProof