egglog_bridge

Module syntax

Egglog proofs reference the source syntax of the query, but the syntax in egglog-bridge is a lower-level, “desugared” representation of that syntax.

This module defines the SourceSyntax and SourceExpr types, which allow callers to reflect the syntax of the original egglog query, along with how it maps to the desugared query language in this crate. The proofs machinery then reconstructs proofs according to this syntax.

Structs

SourceSyntax

A data-structure representing an egglog query. Essentially, multiple SourceExprs, one per line, along with a backing store accounting for subterms indexed by SyntaxId.

SyntaxId

an offset into a Syntax DAG.

Enums

SourceExpr

Representative source syntax for one line of an egglog query, namely, the left-hand-side of an egglog rule.

TopLevelLhsExpr