egglog_bridge

Struct RuleBuilder

pub struct RuleBuilder<'a> { /* private fields */ }

Implementations

impl RuleBuilder<'_>

pub fn egraph(&self) -> &EGraph

Access the underlying egraph within the builder.

pub fn build(self) -> RuleId

Register the given rule with the egraph.

pub fn build_with_syntax(self, syntax: SourceSyntax) -> RuleId

pub fn new_var(&mut self, ty: ColumnTy) -> Variable

Bind a new variable of the given type in the query.

pub fn new_var_named(&mut self, ty: ColumnTy, name: &str) -> QueryEntry

Bind a new variable of the given type in the query.

This method attaches the given name to the QueryEntry, which can make debugging easier.

pub fn call_external_func( &mut self, func: ExternalFunctionId, args: &[QueryEntry], ret_ty: ColumnTy, panic_msg: impl FnOnce() -> String + 'static + Send, ) -> Variable

pub fn query_table( &mut self, func: FunctionId, entries: &[QueryEntry], is_subsumed: Option<bool>, ) -> Result<AtomId>

Add the given table atom to query. As elsewhere in the crate, the last argument is the “return value” of the function. Can also optionally check the subsumption bit.

pub fn query_prim( &mut self, func: ExternalFunctionId, entries: &[QueryEntry], _ret_ty: ColumnTy, ) -> Result<()>

Add the given primitive atom to query. As elsewhere in the crate, the last argument is the “return value” of the function.

pub fn subsume(&mut self, func: FunctionId, entries: &[QueryEntry])

Subsume the given entry in func.

entries should match the number of keys to the function.

pub fn lookup( &mut self, func: FunctionId, entries: &[QueryEntry], panic_msg: impl FnOnce() -> String + Send + 'static, ) -> Variable

Look up the value of a function in the database. If the value is not present, the configured default for the function is used.

For functions configured with DefaultVal::Fail, failing lookups will use panic_msg in the panic output.

pub fn union(&mut self, l: QueryEntry, r: QueryEntry)

Merge the two values in the union-find.

pub fn set(&mut self, func: FunctionId, entries: &[QueryEntry])

Set the value of a function in the database.

pub fn remove(&mut self, table: FunctionId, entries: &[QueryEntry])

Remove the value of a function from the database.

pub fn panic(&mut self, message: String)

Panic with a given message.